[Internal] Refactor exporter: split huge files into smaller ones
This PR splits `context.go` and `util.go` into smaller files, filled with related
functions, like, utils for workspace objects, utils for SCIM (users/sps/groups), ...
