+++ exporter/context_test.go
@@ -378,7 +378,7 @@ func TestGenerateResourceIdForWsObject(t *testing.T) {
 		Importables: resourcesMap,
 		Resources:   p.ResourcesMap,
 	}
-	rid, rtype := ic.generateResourceIdForWsObject(workspace.ObjectStatus{
+	rid, rtype := ic.generateResourceIdForWorkspaceObject(workspace.ObjectStatus{
 		ObjectID:   123,
 		Path:       "Test",
 		ObjectType: "Unknown",
@@ -386,7 +386,7 @@ func TestGenerateResourceIdForWsObject(t *testing.T) {
 	assert.Empty(t, rid)
 	assert.Empty(t, rtype)
 
-	rid, rtype = ic.generateResourceIdForWsObject(workspace.ObjectStatus{
+	rid, rtype = ic.generateResourceIdForWorkspaceObject(workspace.ObjectStatus{
 		ObjectID:   123,
 		Path:       "/Users/user@domain.com/TestDir",
 		ObjectType: workspace.Directory,
@@ -394,7 +394,7 @@ func TestGenerateResourceIdForWsObject(t *testing.T) {
 	assert.Equal(t, "databricks_directory.users_user_domain_com_testdir_123", rid)
 	assert.Equal(t, "databricks_directory", rtype)
 
-	rid, rtype = ic.generateResourceIdForWsObject(workspace.ObjectStatus{
+	rid, rtype = ic.generateResourceIdForWorkspaceObject(workspace.ObjectStatus{
 		ObjectID:   123,
 		Path:       "/Users/user@domain.com/Test File",
 		ObjectType: workspace.File,
@@ -402,7 +402,7 @@ func TestGenerateResourceIdForWsObject(t *testing.T) {
 	assert.Equal(t, "databricks_workspace_file.users_user_domain_com_test_file_123", rid)
 	assert.Equal(t, "databricks_workspace_file", rtype)
 
-	rid, rtype = ic.generateResourceIdForWsObject(workspace.ObjectStatus{
+	rid, rtype = ic.generateResourceIdForWorkspaceObject(workspace.ObjectStatus{
 		ObjectID:   123,
 		Path:       "/Users/user@domain.com/Test Notebook",
 		ObjectType: workspace.Notebook,
